+++ examples/react/src/data-files.tsx
@@ -29,6 +29,7 @@ const toolkit = new Toolkit({
       target: 'pino-pretty',
     },
   }),
+  title: '@arcanejs example',
 });
 
 toolkit.start({