+++ examples/react/src/editable-group-title.tsx
@@ -11,6 +11,7 @@ const toolkit = new Toolkit({
       target: 'pino-pretty',
     },
   }),
+  title: '@arcanejs example',
 });
 
 toolkit.start({