+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # docs
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [3f71c86]
+- Updated dependencies [8038521]
+- Updated dependencies [30aea22]
+  - @arcanejs/toolkit@1.1.0
+  - @arcanejs/protocol@0.4.1
+  - @arcanejs/toolkit-frontend@0.3.1
+
 ## 0.1.15
 
 ### Patch Changes